requirements.txt
google-generativeai
anthropic
openai
mistralai
python-dotenv
colorama